Move fileLoader to common component dir
